Rename Connections to Connect, move under System, fold Logs into it
Consolidates the sidebar's "Integrations" heading (Connections + Logs)
into a single "Connect" entry under the System group. The log viewer
is now a collapsible section on the Connect page, fixed at the bottom
and expanding upward, matching the SystemEvents pattern on the Stats
page, instead of a separate nav item and route.
