+++ frontend/src/App.jsx
@@ -16,7 +16,6 @@ import Settings from './pages/Settings';
 import PluginsPage from './pages/Plugins';
 import PluginBrowsePage from './pages/PluginBrowse';
 import ConnectPage from './pages/Connect';
-import ConnectLogsPage from './pages/ConnectLogs';
 import Users from './pages/Users';
 import LogosPage from './pages/Logos';
 import VODsPage from './pages/VODs';
@@ -161,10 +160,6 @@ const App = () => {
                         />
                         <Route path="/plugins" element={<PluginsPage />} />
                         <Route path="/connect" element={<ConnectPage />} />
-                        <Route
-                          path="/connect/logs"
-                          element={<ConnectLogsPage />}
-                        />
                         <Route path="/users" element={<Users />} />
                         <Route path="/settings" element={<Settings />} />
                         <Route path="/logos" element={<LogosPage />} />
